refactor(gateway): split app and file route groups
